backend/requirements.txt
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
autopep8==1.4.4
colorama==0.4.1
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
importlib-metadata==0.23
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pep8==1.7.1
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
